start packages

+++ database/migrations/2019_08_19_000000_create_failed_jobs_table.php
@@ -27,6 +27,9 @@ return new class extends Migration
      */
     public function down(): void
     {
+        if (!app()->isLocal()) {
+            return;
+        }
         Schema::dropIfExists('failed_jobs');
     }
 };
