+++ database/migrations/2023_11_07_224537_create_attribute_values_table.php
@@ -22,6 +22,9 @@ return new class extends Migration {
 
 	public function down()
 	{
+        if (!app()->isLocal()) {
+            return;
+        }
 		Schema::dropIfExists('attribute_values');
 	}
 };
