+++ database/migrations/2023_11_18_130446_create_sliders_table.php
@@ -22,6 +22,9 @@ return new class extends Migration {
 
     public function down(): void
     {
+        if (!app()->isLocal()) {
+            return;
+        }
         Schema::dropIfExists('sliders');
     }
 };
